refactor move ChartDataResult enums to common (#17399)
